rustc, rustc_passes: don't depend on syntax_expand.
This is done by moving some data definitions to syntax::expand.
